WIP: refactor: making the FilesChecksum agnostic
