+++ pydatacuration/utils/files_checksum.py
@@ -18,15 +18,14 @@ class FilesChecksum:
         """Generate the checksum of the files in the dataset directory.
 
         Args:
-            target_dir (Path): The path to the dataset directory.
+            target_dir (Path): The path to directory containing the files.
 
         Returns:
             list: A list of dictionaries containing the file path and the checksum.
         """
         dl_file_checksum_nested_list = []
 
-        # Join the workdir with 'dataset' and 'files' to get the target directory, and resolve to an absolute path
-        target_dir_path = Path(target_dir, 'dataset', 'files').resolve()
+        target_dir_path = target_dir.resolve()
 
         # Iterate through all files in the directory and subdirectories
         for file_path in target_dir_path.rglob('*'):
